+++ packages/core/Cargo.toml
@@ -13,6 +13,7 @@ release = false
 [dependencies]
 base64 = { version = "0.22.0", default-features = false, features = ["alloc"] }
 bnum = "0.11.0"
+cosmwasm-crypto = { version = "2.0.0", path = "../crypto" }
 crc32fast = { version = "1.4.0", optional = true, default-features = false }
 derive_more = { version = "1.0.0-beta.6", default-features = false, features = ["display", "from"] }
 hex = { version = "0.4.3", default-features = false, features = ["alloc"] }
