+++ packages/core/src/errors/mod.rs
@@ -1,11 +1,16 @@
+mod backtrace;
 mod core_error;
+mod recover_pubkey_error;
+mod system_error;
+mod verification_error;
 
-pub(crate) mod backtrace;
-
-pub type CoreResult<T, E = CoreError> = core::result::Result<T, E>;
-
+pub(crate) use backtrace::{impl_from_err, BT};
 pub use core_error::{
     CheckedFromRatioError, CheckedMultiplyFractionError, CheckedMultiplyRatioError,
-    ConversionOverflowError, CoreError, DivideByZeroError, DivisionError, OverflowError,
-    OverflowOperation, RoundDownOverflowError, RoundUpOverflowError,
+    CoinFromStrError, CoinsError, ConversionOverflowError, CoreError, CoreResult,
+    DivideByZeroError, DivisionError, OverflowError, OverflowOperation, RoundDownOverflowError,
+    RoundUpOverflowError,
 };
+pub use recover_pubkey_error::RecoverPubkeyError;
+pub use system_error::SystemError;
+pub use verification_error::VerificationError;
