+++ packages/core/src/errors/recover_pubkey_error.rs
@@ -1,19 +1,19 @@
 use core::fmt::Debug;
 #[cfg(not(target_arch = "wasm32"))]
 use cosmwasm_crypto::CryptoError;
+use derive_more::Display;
 
 use super::BT;
-use thiserror::Error;
 
-#[derive(Error, Debug)]
+#[derive(Display, Debug)]
 pub enum RecoverPubkeyError {
-    #[error("Invalid hash format")]
+    #[display("Invalid hash format")]
     InvalidHashFormat,
-    #[error("Invalid signature format")]
+    #[display("Invalid signature format")]
     InvalidSignatureFormat,
-    #[error("Invalid recovery parameter. Supported values: 0 and 1.")]
+    #[display("Invalid recovery parameter. Supported values: 0 and 1.")]
     InvalidRecoveryParam,
-    #[error("Unknown error: {error_code}")]
+    #[display("Unknown error: {error_code}")]
     UnknownErr { error_code: u32, backtrace: BT },
 }
 
@@ -26,6 +26,9 @@ impl RecoverPubkeyError {
         }
     }
 }
+
+#[cfg(feature = "std")]
+impl std::error::Error for RecoverPubkeyError {}
 
 impl PartialEq<RecoverPubkeyError> for RecoverPubkeyError {
     fn eq(&self, rhs: &RecoverPubkeyError) -> bool {
