--- packages/std/src/errors/mod.rs
@@ -1,14 +1,0 @@
-mod recover_pubkey_error;
-mod std_error;
-mod system_error;
-mod verification_error;
-
-pub(crate) use cosmwasm_core::__internal::{impl_from_err, BT};
-pub use cosmwasm_core::{
-    CheckedFromRatioError, CheckedMultiplyFractionError, CheckedMultiplyRatioError,
-    ConversionOverflowError, DivideByZeroError, DivisionError, OverflowError, OverflowOperation,
-};
-pub use recover_pubkey_error::RecoverPubkeyError;
-pub use std_error::{CoinFromStrError, CoinsError, StdError, StdResult};
-pub use system_error::SystemError;
-pub use verification_error::VerificationError;
